- Fixed sharpmake behavior in filter mode for submit assistant when everything is filtered. - Added File.Exists test in TryToDeleteFile() helper.
(cherry picked from commit 39334620409bd4bc5a89c0af12dcc920bd78b75b)

+++ Sharpmake/Util.cs
@@ -1069,7 +1069,8 @@ namespace Sharpmake
         {
             try
             {
-                File.Delete(filename);
+                if (File.Exists(filename))
+                    File.Delete(filename);
                 return true;
             }
             catch (Exception)
